Fix: Add dtype conversion when writing KV to cache in both SDPA functions
- Add k and v parameters to rotary_kv_update_sdpa_gen_gptoss signature
- Apply RoPE to k in GPT-OSS DECODE_SINGLE path (was missing)
- Implement KV cache write logic in GPT-OSS path (was missing, causing gibberish)
- Convert k and v to cache dtype before writing (fixes bf16->float32 mismatch)
- Also fix standard rotary_kv_update_sdpa_gen for consistency
- Update call site to pass k and v tensors to GPT-OSS function

+++ yalis/attention/sdpa_and_flex.py
@@ -139,13 +139,13 @@ def rotary_kv_update_sdpa_gen(
     if hasattr(ax.config, 'G_intra_c') and ax.config.G_intra_c > 1 and use_intra_head_parallelism:
         k_cache[b_indices, :, t_indices, :] = Drop.apply(
             k[:, :, 0, :], ax.comm_handle.inner_intra_layer_parallel_group
-        )
+        ).to(k_cache.dtype)
         v_cache[b_indices, :, t_indices, :] = Drop.apply(
             v[:, :, 0, :], ax.comm_handle.inner_intra_layer_parallel_group
-        )
+        ).to(v_cache.dtype)
     else:
-        k_cache[b_indices, :, t_indices, :] = k[:, :, 0, :]
-        v_cache[b_indices, :, t_indices, :] = v[:, :, 0, :]
+        k_cache[b_indices, :, t_indices, :] = k[:, :, 0, :].to(k_cache.dtype)
+        v_cache[b_indices, :, t_indices, :] = v[:, :, 0, :].to(v_cache.dtype)
     # Build causal mask (allows keys up to the current index)
     # Optionally constrain to a fixed sliding window if `sliding_window > 0`.
     t_max = k_cache.size(-2)
@@ -283,13 +283,13 @@ def rotary_kv_update_sdpa_gen_gptoss(
     if use_intra_head_parallelism:
         k_cache[b_indices, :, t_indices, :] = Drop.apply(
             k[:, :, 0, :], process_group
-        )
+        ).to(k_cache.dtype)
         v_cache[b_indices, :, t_indices, :] = Drop.apply(
             v[:, :, 0, :], process_group
-        )
+        ).to(v_cache.dtype)
     else:
-        k_cache[b_indices, :, t_indices, :] = k[:, :, 0, :]
-        v_cache[b_indices, :, t_indices, :] = v[:, :, 0, :]
+        k_cache[b_indices, :, t_indices, :] = k[:, :, 0, :].to(k_cache.dtype)
+        v_cache[b_indices, :, t_indices, :] = v[:, :, 0, :].to(v_cache.dtype)
 
     # build mask up to current token_counter and optional sliding window
     t_max = k_cache.size(-2)
